requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
coverage==7.10.7
exceptiongroup==1.3.1
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.0
requests==2.31.0
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
zipp==3.23.0
